requirements.txt
django
djangorestframework
markdown 
django-filter
drf-yasg
mysqlclient